Created and used Cell class
Previously for every new frame I was creating a new rectangle and adding it in the scene
and removing the previous rectangle, which should be more resources heavy. Now I create and
add rectangle to the scene for the first frame and use the same in next frames.

I Created a cell class, so that I can create an 2D array of cells and operate on the same.